Updated nrf-hal
Also, a blinky example that exercises the UART given recent changes there with nrf-hal. The example is consistent with the nRF9160-dk bsp project.

The dependency on the nrf-hal is changed to leverage master. We could probably merge this commit, but withhold a release until nrf-hal is once again released.